refactor: router 분리 및 index 코드 App 으로 이동
